+++ assembly/dependencies-apache-ignite-lgpl.xml
@@ -137,6 +137,7 @@
                 <exclude>${project.groupId}:ignite-sqlline</exclude>
                 <exclude>${project.groupId}:ignite-control-utility</exclude>
                 <exclude>${project.groupId}:ignite-checkstyle</exclude>
+                <exclude>${project.groupId}:ignite-grid-unsafe</exclude>
             </excludes>
             <sources>
                 <includeModuleDirectory>true</includeModuleDirectory>
