+++ modules/calcite/pom.xml
@@ -72,7 +72,6 @@
         <dependency>
             <groupId>${project.groupId}</groupId>
             <artifactId>ignite-grid-unsafe</artifactId>
-            <version>${revision}</version>
             <scope>provided</scope>
         </dependency>
