+++ modules/compatibility/pom.xml
@@ -49,7 +49,12 @@
         <dependency>
             <groupId>${project.groupId}</groupId>
             <artifactId>ignite-commons</artifactId>
-            <version>${revision}</version>
+            <scope>provided</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>ignite-binary-api</artifactId>
             <scope>provided</scope>
         </dependency>
