+++ modules/control-utility/pom.xml
@@ -56,7 +56,6 @@
         <dependency>
             <groupId>${project.groupId}</groupId>
             <artifactId>ignite-grid-unsafe</artifactId>
-            <version>${revision}</version>
             <scope>provided</scope>
         </dependency>
