+++ modules/core/pom.xml
@@ -70,7 +70,6 @@
         <dependency>
             <groupId>${project.groupId}</groupId>
             <artifactId>ignite-grid-unsafe</artifactId>
-            <version>${revision}</version>
             <scope>provided</scope>
         </dependency>
