+++ modules/direct-io/pom.xml
@@ -51,7 +51,6 @@
         <dependency>
             <groupId>${project.groupId}</groupId>
             <artifactId>ignite-grid-unsafe</artifactId>
-            <version>${revision}</version>
             <scope>provided</scope>
         </dependency>
