+++ modules/indexing/pom.xml
@@ -55,7 +55,6 @@
         <dependency>
             <groupId>${project.groupId}</groupId>
             <artifactId>ignite-grid-unsafe</artifactId>
-            <version>${revision}</version>
             <scope>provided</scope>
         </dependency>
